hedera-node/test-clients/scripts/diff-testing/requirements.txt
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.2
Deprecated==1.2.14
docstring-parser==0.15
google-api-core==2.16.2
google-auth==2.27.0
google-cloud-core==2.4.1
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
idna==3.6
jsons==1.6.3
mypy-extensions==1.0.0
protobuf==4.25.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
PyNaCl==1.5.0
requests==2.31.0
rsa==4.9
typed-argument-parser==1.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
typish==1.9.3
urllib3==2.2.0
wrapt==1.16.0
